+++ src/Planners/include/BipedalLocomotion/Planners/CubicSpline.h
@@ -1,12 +1,12 @@
 /**
- * @file QuinticSpline.h
+ * @file CubicSpline.h
  * @authors Giulio Romualdi
- * @copyright 2020 Istituto Italiano di Tecnologia (IIT). This software may be modified and
+ * @copyright 2021 Istituto Italiano di Tecnologia (IIT). This software may be modified and
  * distributed under the terms of the GNU Lesser General Public License v2.1 or any later version.
  */
 
-#ifndef BIPEDAL_LOCOMOTION_PLANNERS_QUINTIC_SPLINE_H
-#define BIPEDAL_LOCOMOTION_PLANNERS_QUINTIC_SPLINE_H
+#ifndef BIPEDAL_LOCOMOTION_PLANNERS_CUBIC_SPLINE_H
+#define BIPEDAL_LOCOMOTION_PLANNERS_CUBIC_SPLINE_H
 
 #include <memory>
 #include <vector>
@@ -20,9 +20,9 @@ namespace BipedalLocomotion
 namespace Planners
 {
 /**
- * Quintic spline implement a 5-th order polynomial spline in \$f\mathbb{R}^n\$f.
+ * Cubic spline implement a 3-rd order polynomial spline in \$f\mathbb{R}^n\$f.
  */
-class QuinticSpline : public Spline
+class CubicSpline : public Spline
 {
     /**
      * Private implementation of the class
@@ -34,13 +34,13 @@ public:
     /**
      * Constructor.
      */
-    QuinticSpline();
+    CubicSpline();
 
     /**
      * Destructor.
      * @note It is required by the PIMPL idiom.
      */
-    ~QuinticSpline();
+    ~CubicSpline();
 
     /**
      * Set the time step of the advance interface.
@@ -63,6 +63,7 @@ public:
      * Set the initial condition of the spline
      * @param velocity initial velocity (i.e. first derivative).
      * @param acceleration initial acceleration (i.e. second derivative).
+     * @note the acceleration is not considered in the spline evaluation
      * @return True in case of success, false otherwise.
      */
     bool setInitialConditions(Eigen::Ref<const Eigen::VectorXd> velocity,
@@ -72,6 +73,7 @@ public:
      * Set the final condition of the spline
      * @param velocity final velocity (i.e. first derivative).
      * @param acceleration final acceleration (i.e. second derivative).
+     * @note the acceleration is not considered in the spline evaluation
      * @return True in case of success, false otherwise.
      */
     bool setFinalConditions(Eigen::Ref<const Eigen::VectorXd> velocity,
@@ -125,4 +127,4 @@ public:
 } // namespace Planners
 } // namespace BipedalLocomotion
 
-#endif // BIPEDAL_LOCOMOTION_PLANNERS_QUINTIC_SPLINE_H
+#endif // BIPEDAL_LOCOMOTION_PLANNERS_CUBIC_SPLINE_H